KS-1-4 metrics of Gagarin's flight with fixions

+++ src/main/kotlin/less_1/less1_task4.kt
@@ -6,7 +6,7 @@ The metrics of Gagarin's flight
 fun main() {
     val lenght: Long = 40868600000
     val age: Byte = 27
-    val partOfDay: Double = 0.075
+    val partOfDay: Float = 0.075f
     val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
